requirements.txt
art==6.1
Pillow==10.0.0
rich==13.6.0